calendar - separated component
