+++ src/components/User/UserStatistics/UserStatistics.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import {
   BarChart,
   Bar,
@@ -9,30 +9,9 @@ import {
   CartesianGrid,
 } from 'recharts';
 
-import DatePicker from "react-datepicker";
-
-import "react-datepicker/dist/react-datepicker.css";
-
 import css from './UserStatistics.module.css';
 
-
-export const CalendarDropdown = () => {
-    const [date, setDate] = useState(new Date());
-
-    const handleCalendarClose = () => console.log("Calendar closed");
-    const handleCalendarOpen = () => console.log("Calendar opened");
-  
-    return (
-      <DatePicker
-        selected={date}
-        onChange={(date) => setDate(date)}
-        onCalendarClose={handleCalendarClose}
-        onCalendarOpen={handleCalendarOpen}
-      />
-    );
-  };
-
-//data for stats chart 
+//data for stats chart
 const data = [
   { name: 'To Do', month: 100, day: 60, amt: 100 },
   { name: 'In Progress', month: 100, day: 40, amt: 100 },
