+++ src/pages/Homepage.js
@@ -1,7 +1,8 @@
 import { Description } from 'components/Homepage/DescSection/DescSection';
 import { HeroSection } from 'components/Homepage/HeroSection/HeroSection';
 import { ReviewsSlider, TitleH2 } from 'components/Homepage/ReviewsSection/ReviewsSection';
-import { CalendarDropdown, StatisticsChart } from 'components/User/UserStatistics/UserStatistics';
+import { CalendarDropdown } from 'components/User/UserCalendar/UserCalendar';
+import { StatisticsChart } from 'components/User/UserStatistics/UserStatistics';
 
 const Homepage = () => {
   return (
